+++ src/nserver/__init__.py
@@ -1,5 +1,5 @@
 from .models import Query, Response
-from .rules import RegexRule, WildcardStringRule
+from .rules import StaticRule, ZoneRule, RegexRule, WildcardStringRule
 from .records import A, AAAA, NS, CNAME, PTR, SOA, MX, TXT, CAA
 from .server import NameServer
 from .settings import Settings